% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ug.R
\name{calcUg}
\alias{calcUg}
\title{Calculate the zonal group velocity}
\usage{
calcUg(betamz, Ks2, y, lat, K = 3, a = 6371000)
}
\arguments{
\item{betamz}{meridional gradient of the absolute vorticity
in Mercator coordinates}

\item{Ks2}{Numeric. Stationary Rossby wavenumber}

\item{y}{A latitude}

\item{lat}{Numeric vector of latitudes from minor to major}

\item{K}{Numeric. Total Rossby wavenumber}

\item{a}{Earth's radio}
}
\value{
Numeric.
}
\description{
This function calculates the zonal group velocity.
}
\examples{
{
input <- system.file("extdata",
                     "uwnd.mon.mean_200hPa_2014JFM.nc",
                      package = "raytracing")
b <- betaks(ifile = input)
betam_z <- colMeans(b$betam, na.rm = TRUE)
um_z <- colMeans(b$um, na.rm = TRUE)
y0 <- -30
calcUg(betamz = betam_z,
       Ks2 = betam_z[ypos(y0, b$lat)]/um_z[ypos(y0, b$lat)],
       y = y0,
       lat = b$lat) # 7.039724e-05
}
}
