% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vg.R
\name{calcVg}
\alias{calcVg}
\title{Calculate the meridional group velocity}
\usage{
calcVg(betamz, umz, y, lat, K = 3, direction = -1, tl = 1, a = 6371000)
}
\arguments{
\item{betamz}{meridional gradient of the absolute vorticity
in Mercator coordinates}

\item{umz}{zonal wind in mercator coordinates}

\item{y}{A latitude}

\item{lat}{numeric vector of latitudes from minor to major}

\item{K}{Total Rossby wavenumber}

\item{direction}{Controls the wave displacement.
The wave goes to the north of the source using 1 and
to the south of the source using -1.}

\item{tl}{Turning latitude. It will always start with a
positive tl and automatically change to negative
after the turning latitude or before a critical value.}

\item{a}{Earth's radio}
}
\value{
The meridional group velocity
}
\description{
This function calculates the meridional group velocity.
}
\examples{
{
input <- system.file("extdata",
                     "uwnd.mon.mean_200hPa_2014JFM.nc",
                      package = "raytracing")
b <- betaks(ifile = input)
y0 <- -30
calcVg(betamz = colMeans(b$betam, na.rm = TRUE),
       umz = colMeans(b$um, na.rm = TRUE),
       y = y0,
       lat = b$lat)
}
}
