% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coastlines.R
\docType{data}
\name{coastlines}
\alias{coastlines}
\title{Coastlines}
\format{
Geometry of coastlines "sfc_MULTILINESTRING"
\describe{
\item{}{Geometry of coastlines "sfc_MULTILINESTRING"}
data(coastlines)
}
}
\source{
\url{https://www.naturalearthdata.com/downloads/10m-physical-vectors/10m-coastline/}
}
\usage{
data(coastlines)
}
\description{
Geometry of coastlines, class "sfc_MULTILINESTRING" "sfc" from the
package "sf"
}
\keyword{datasets}
