% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigon_interpolation.R
\name{it}
\alias{it}
\title{Trigonometric interpolation}
\usage{
it(y, yk, nlat)
}
\arguments{
\item{y}{Numeric. The latitude where the interpolation is
required}

\item{yk}{Numeric vector of the data to be interpolated.
For instance, umz or betam}

\item{nlat}{Numeric. The total latitudes of the vector yk.}

\item{betamz}{meridional gradient of the absolute vorticity
in Mercator coordinates}
}
\value{
Numeric value
}
\description{
This function performs trigonometric interpolation for the
passed basic state variable and the requested latitude
}
\examples{
{
input <- system.file("extdata",
                     "uwnd.mon.mean_200hPa_2014JFM.nc",
                      package = "raytracing")
b <- betaks(ifile = input)
umz <- colMeans(b$um, na.rm = TRUE)
y0 <- -17
it(y = -17, yk = umz, nlat = length(b$lat))
}
}
