% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ray.R
\name{ray}
\alias{ray}
\title{Calculates the Rossby waves ray paths}
\usage{
ray(
  betamz,
  umz,
  lat,
  x0,
  y0,
  K,
  dt,
  itime,
  direction,
  tl = 1,
  a = 6371000,
  verbose = FALSE,
  ofile
)
}
\arguments{
\item{betamz}{zonal mean of \strong{betam}. \strong{betam} is a
matrix (longitude = rows x latitude from minor to major = columns)
obtained with \code{\link{betaks}}. It is suggested to use
\code{\link{colMeans}} base R function in order to obtain the
zonal mean of \strong{betam}.}

\item{umz}{zonal mean of \strong{um}. \strong{um} is a
matrix (longitude = rows x latitude from minor to major = columns)
obtained with \code{\link{betaks}}. It is suggested to use
\code{\link{colMeans}} base R function in order to obtain the
zonal mean of \strong{um}.}

\item{lat}{Numeric vector of latitudes from minor to major
(ex: -90 to 90)}

\item{x0}{Numeric value; longitude (0 to 360 degrees)}

\item{y0}{Numeric value; latitude}

\item{K}{Numeric value; Total Rossby wavenumber}

\item{dt}{Numeric value; Timestep for integration (seconds)}

\item{itime}{Numeric value; total integration time. For instance, 10 days
times 4 times per day.}

\item{direction}{Numeric value (possibilities: 1 or -1).
It controls the wave displacement:
If 1, the wave goes to the north of the source;
If -1, the wave goes to the south of the source.}

\item{tl}{Numeric value; Turning latitude. Do not change this!
It will always start with a positive tl (1) and automatically
change to negative (-1) after the turning latitude.}

\item{a}{Earth's radio (m)}

\item{verbose}{Boolean; if TRUE (default) return messages
during compilation}

\item{ofile}{Character;  Output file name with .csv extension,
for instance, "/user/ray.csv"}
}
\value{
sf data.frame
}
\description{
\code{ray} must ingest the zonal means of \strong{betam} and \strong{um},
plus the latitude vector (\strong{lat}). \code{ray} returns the Rossby wave
ray paths (lat/lon) triggered from an initial source/position (x0, y0),
a total wavenumber (K), and direction set up when invoking the function.
}
\examples{
{
# For Magana and Ambrizzi (1998):
input <- system.file("extdata",
                     "uwnd.mon.mean_300hPa_1997-98DJF.nc",
                      package = "raytracing")
b <- betaks(ifile = input)
ma98 <- ray(betamz = colMeans(b$betam, na.rm = TRUE),
                umz = colMeans(b$um, na.rm = TRUE),
                lat = b$lat,
                K = 3,
                itime = 10*4,
                x0 = -130 + 360,
                y0 = -17,
                dt = 6 * 60 * 60,
                direction = -1)
head(ma98)
# For Coelho et al. (2015):
input <- system.file("extdata",
                     "uwnd.mon.mean_200hPa_2014JFM.nc",
                      package = "raytracing")
b <- betaks(ifile = input)
co2015 <- ray(betamz = colMeans(b$betam, na.rm = TRUE),
                umz = colMeans(b$um, na.rm = TRUE),
                lat = b$lat,
                K = 3,
                itime = 10 * 4,
                x0 = -130 + 360,
                y0 = -30,
                dt = 6 * 60 * 60,
                direction = -1)
head(co2015)
}
}
