% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ray_path.R
\name{ray_path}
\alias{ray_path}
\title{Calculate the ray paths / great circles}
\usage{
ray_path(x, y)
}
\arguments{
\item{x}{vector with the longitude obtained with \code{ray} or \code{ray_source}}

\item{y}{vector with the latitude obtained with \code{ray} or \code{ray_source}}
}
\value{
sf data.frame
}
\description{
This function calculates the great circles using the (x0, y0) obtained
with \code{ray} or \code{ray_source}. It returns a LINESTRING geometry that is ready
for plot.
}
\examples{
{
# For Magana and Ambrizzi (1998):
input <- system.file("extdata",
                     "uwnd.mon.mean_300hPa_1997-98DJF.nc",
                      package = "raytracing")
b <- betaks(ifile = input)
ma98 <- ray(betamz = colMeans(b$betam, na.rm = TRUE),
                umz = colMeans(b$um, na.rm = TRUE),
                lat = b$lat,
                K = 3,
                itime = 10,
                x0 = -130,
                y0 = -17,
                dt = 12 * 60 * 60,
                direction = -1)

# For Coelho et al. (2015):
input <- system.file("extdata",
                     "uwnd.mon.mean_200hPa_2014JFM.nc",
                      package = "raytracing")
b <- betaks(ifile = input)
co2015 <- ray(betamz = colMeans(b$betam, na.rm = TRUE),
          umz = colMeans(b$um, na.rm = TRUE),
          lat = b$lat,
          K = 3,
          itime = 30,
          x0 = -135 + 360,
          y0 = -30,
          dt = 6 * 60 * 60,
          direction = -1)

r1 <- co2015
# We want to obtain all the linestrings except for the turning points
rp <- ray_path(x = r1$x0, y = r1$y0)
plot(rp, axes  = TRUE, col = "red")


### ray source
r1 <- ray_source(betamz = colMeans(b$betam, na.rm = TRUE),
                umz = colMeans(b$um, na.rm = TRUE),
                lat = b$lat,
                direction = -1,
                x0 = -(130) + 360,
                y0 = -(17),
                K = c(3, 4, 5),
                dt = 12 * 60 * 60,
                itime = 30)
}
}
