% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ray_source.R
\name{ray_source}
\alias{ray_source}
\title{Calculate the Rossby waves ray paths over a source region}
\usage{
ray_source(
  betamz,
  umz,
  lat,
  x0,
  y0,
  K,
  dt,
  itime,
  direction,
  tl = 1,
  a = 6371000,
  verbose = FALSE,
  ofile
)
}
\arguments{
\item{betamz}{zonal mean of \strong{betam}. \strong{betam} is a
matrix (longitude = rows x latitude from minor to major = columns)
obtained with \code{\link{betaks}}. It is suggested to use
\code{\link{colMeans}} base R function in order to obtain the
zonal mean of \strong{betam}.}

\item{umz}{zonal mean of \strong{um}. \strong{um} is a
matrix (longitude = rows x latitude from minor to major = columns)
obtained with \code{\link{betaks}}. It is suggested to use
\code{\link{colMeans}} base R function in order to obtain the
zonal mean of \strong{um}.}

\item{lat}{Numeric vector of latitudes from minor to major
(ex: -90 to 90)}

\item{x0}{Vector with longitudes (0 to 360 degrees)}

\item{y0}{Vector with latitudes}

\item{K}{Vector; Total Rossby wavenumber}

\item{dt}{Numeric value; Timestep for integration (seconds)}

\item{itime}{Numeric value; total integration time}

\item{direction}{Vector with two possibilities: 1 or -1
It controls the wave displacement:
If 1, the wave goes to the north of the source;
If -1, the wave goes to the south of the source.}

\item{tl}{Numeric value; Turning latitude. Do not change this!
It will always start with a positive tl (1) and automatically
change to negative (-1) after the turning latitude.}

\item{a}{Earth's radio (m)}

\item{verbose}{Boolean; if TRUE (default) return messages
during compilation}

\item{ofile}{Character;  Output file name with .csv extension,
for instance, "/user/ray.csv"}
}
\value{
sf data.frame
}
\description{
\code{ray_source} must ingest the zonal means of \strong{betam} and
\strong{um}, plus the latitude vector (\strong{lat}). \code{ray_source}
returns the Rossby wave ray paths (lat/lon) triggered according a set of
experiments. The ray paths will be calculated for the combination
of initial sources/positions (x0, y0), total wavenumbers (K),
and directions, using the same basic state given by  \code{\link{betaks}}.
}
\examples{
{
input <- system.file("extdata",
                     "uwnd.mon.mean_300hPa_1997-98DJF.nc",
                      package = "raytracing")
b <- betaks(ifile = input)
r1 <- ray_source(betamz = colMeans(b$betam, na.rm = TRUE),
                umz = colMeans(b$um, na.rm = TRUE),
                lat = b$lat,
                direction = -1,
                x0 = -130 + 360,
                y0 = -17,
                K = c(5),
                dt = 12 * 60 * 60,
                itime = 10*2)
# Simple plot:

}
}
