% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raytracing-package.R
\docType{package}
\name{raytracing-package}
\alias{raytracing}
\alias{raytracing-package}
\title{raytracing: Rossby wave ray tracing}
\description{
Calculates the atmospheric Rossby wave ray tracks using 
             zonal symetric flow basic state based on Hoskins and Karoly 
             (1981) <doi:10.1175/1520-0469(1981)038<1179:TSLROA>2.0.CO;2>,
             Karoly (1983) <doi:10.1016/0377-0265(83)90013-1>, 
             Hoskins and Ambrizzi (1993)
             <doi:10.1175/1520-0469(1993)050<1661:RWPOAR>2.0.CO;2>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/salvatirehbein/raytracing/}
  \item Report bugs at \url{https://github.com/salvatirehbein/raytracing/issues/}
}

}
\author{
\strong{Maintainer}: Amanda Rehbein \email{amanda.rehbein@usp.br} (\href{https://orcid.org/0000-0002-8714-7931}{ORCID})

Other contributors:
\itemize{
  \item Sergio Ibarra-Espinosa \email{sergio.ibarra@usp.br} (\href{https://orcid.org/0000-0002-3162-1905}{ORCID}) [contributor]
}

}
\keyword{internal}
