% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ypos.R
\name{ypos}
\alias{ypos}
\title{Get the position of a given latitute y}
\usage{
ypos(y, lat)
}
\arguments{
\item{y}{numeric value of one latitude}

\item{lat}{numeric vector of latitudes from minor to major}
}
\value{
The position where the latitude y has the minor
difference with lat
}
\description{
This function get the position in a vector of a given latitute y.
}
\examples{
{
ypos(y = -30, lat = seq(-90, 90, 2.5))
}
}
