% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_ks.R
\name{betaks}
\alias{betaks}
\title{Calculates Beta and Ks}
\usage{
betaks(
  ifile,
  varname = "uwnd",
  latname = "lat",
  lonname = "lon",
  ofile,
  a = 6371000,
  plots = FALSE,
  show.warnings = FALSE
)
}
\arguments{
\item{ifile}{String indicating the input data filename. The file to be
passed consists in a netCDF file with only time-mean zonal wind at one
pressure level and with latitude in ascending order (not a requisite).
It is required that the read dimensions express
longitude (in rows) x latitude (in columns)}

\item{varname}{String indicating the Variable name of the time-mean
zonal wind field}

\item{latname}{String indicating the variable name of the latitude field}

\item{lonname}{String indicating the variable name of the longitude field}

\item{ofile}{String indicating the file name for store output data.
If missing, will not return a netCDF file.}

\item{a}{Numeric indicating the Earth's radio (m)}

\item{plots}{Logical, if TRUE will produce filled.countour plots}

\item{show.warnings}{Logical, if TRUE will warns about NaNs in sqrt(<0)}
}
\value{
list with one vector (lat) and 3 matrices (um, betam, and ksm)
}
\description{
\code{betaks} ingests the time-mean zonal wind (u), transform it in
mercator coordinates (um); calculates the meridional gradient of
the absolute vorticity (beta) in mercator coordinates (betam);
and, finally, calculates stationary wavenumber (Ks) in mercator coordinates
(ksm) (see: Hoskins and Ambrizzi, 1993). \code{betaks} returns the um, betam,
and lat, for being ingested in \code{\link{ray}} or
\code{\link{ray_source}}.
}
\examples{
{
input <- system.file("extdata",
                     "uwnd.mon.mean_200hPa_2014JFM.nc",
                      package = "raytracing")
b <- betaks(ifile = input)
output <- paste0(tempfile(), ".nc")
b <- betaks(ifile = input, ofile = output)
}
}
