% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_ks.R
\name{betaks}
\alias{betaks}
\title{Calculate the meridional gradient of the absolute vorticity (beta)
and the stationary Rossby wavenumber (Ks)}
\usage{
betaks(
  ifile,
  varname = "uwnd",
  latname = "lat",
  lonname = "lon",
  ofile,
  a = 6371000,
  plots = FALSE,
  show.warnings = FALSE
)
}
\arguments{
\item{ifile}{Input file name consisting in zonal wind at one pressure
level and one time}

\item{varname}{Zonal wind variable name at a specified level and time}

\item{latname}{Latitude variable name}

\item{lonname}{Longitude variable name}

\item{ofile}{Output file name}

\item{a}{Earth's radio (m)}

\item{plots}{Logical, if TRUE will produce filled.countour plots}

\item{show.warnings}{Logical, if TRUE will warns about NaNs in sqrt(<0)}
}
\value{
The meridional group velocity
}
\description{
This function calculates the meridional gradient of the absolute
vorticity (beta), zonal wind (u), and stationary Rossby wave in
mercator coordinates as in Hoskins & Ambrizzi (1993).
The input data is zonal wind with latitude in ascending order.
}
\examples{
{
input <- system.file("extdata",
                     "uwnd.mon.mean_200hPa_2014JFM.nc",
                      package = "raytracing")
b <- betaks(ifile = input)
output <- paste0(tempfile(), ".nc")
b <- betaks(ifile = input, ofile = output)
}
}
