% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ug.R
\name{calcUg}
\alias{calcUg}
\title{Calculate the zonal group velocity}
\usage{
calcUg(betamz, umz, y, lat, K = 3, w = 0, a = 6371000)
}
\arguments{
\item{betamz}{meridional gradient of the absolute vorticity
in Mercator coordinates}

\item{umz}{zonal wind in mercator coordinates}

\item{y}{A latitude}

\item{lat}{numeric vector of latitudes from minor to major}

\item{K}{Total Rossby wavenumber}

\item{w}{wave frequency}

\item{a}{Earth's radio}
}
\value{
The zonal group velocity
}
\description{
This function calculates the zonal group velocity.
}
\examples{
{
input <- system.file("extdata",
                     "uwnd.mon.mean_200hPa_2014JFM.nc",
                      package = "raytracing")
b <- betaks(ifile = input)
y0 <- -30
calcUg(betamz = colMeans(b$betam, na.rm = TRUE),
       umz = colMeans(b$um, na.rm = TRUE),
       y = y0,
       lat = b$lat) # 7.039724e-05
}
}
