% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ug.R
\name{calcUg}
\alias{calcUg}
\title{Calculate the zonal group velocity}
\usage{
calcUg(beta, Ks2, lat, K = 3, a = 6371000)
}
\arguments{
\item{beta}{meridional gradient of the absolute vorticity
in Mercator coordinates}

\item{Ks2}{Numeric. Stationary Rossby wavenumber}

\item{lat}{Numeric vector of latitudes from minor to major}

\item{K}{Numeric. Total Rossby wavenumber}

\item{a}{Earth's radio}
}
\value{
Numeric.
}
\description{
This function calculates the zonal group velocity.
}
\examples{
{
input <- system.file("extdata",
                     "uwnd.mon.mean_200hPa_2014JFM.nc",
                      package = "raytracing")
b <- betaks(u = input)
y0 <- -30
lat <- rev(b$lat)
phirad <- lat*pi/180
betamz <- rev(colMeans(b$betam, na.rm = TRUE))
umz <- rev(colMeans(b$u, na.rm = TRUE))*cos(phirad)
beta_y0 <- trin(y0, yk = betamz)
u_y0 <- trin(y0, yk = umz)
Ks2_y0 <- beta_y0/u_y0

calcUg(beta = beta_y0,
       Ks2 = Ks2_y0,
       lat = lat)
}
}
