% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vg.R
\name{calcVg}
\alias{calcVg}
\title{Calculate the meridional group velocity}
\usage{
calcVg(beta, Ks2, y, lat, K = 3, direction = -1, tl = 1, a = 6371000)
}
\arguments{
\item{beta}{meridional gradient of the absolute vorticity
in Mercator coordinates}

\item{Ks2}{Numeric. Stationary Rossby wavenumber}

\item{y}{The latitude y0 or y1}

\item{lat}{Numeric vector of latitudes from minor to major}

\item{K}{Numeric. Total Rossby wavenumber}

\item{direction}{Controls the wave displacement.
The wave goes to the north of the source using 1 and
to the south of the source using -1.}

\item{tl}{Numeric. Turning latitude. It will always start with a
positive tl and automatically change to negative
after the turning latitude or before a critical value.}

\item{a}{Earth's radio}
}
\value{
Numeric.
}
\description{
This function calculates the meridional group velocity.
}
\examples{
{
input <- system.file("extdata",
                     "uwnd.mon.mean_200hPa_2014JFM.nc",
                      package = "raytracing")
b <- betaks(u = input)
y0 <- -30
lat <- rev(b$lat)
phirad <- lat*pi/180
betamz <- rev(colMeans(b$betam, na.rm = TRUE))
umz <- rev(colMeans(b$u, na.rm = TRUE))*cos(phirad)
beta_y0 <- trin(y0, yk = betamz)
u_y0 <- trin(y0, yk = umz)
Ks2_y0 <- beta_y0/u_y0

calcVg(beta = beta_y0,
       Ks2 = Ks2_y0,
       y = y0,
       lat = lat)
}
}
