% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ray.R
\name{ray}
\alias{ray}
\title{Calculate the Rossby waves ray paths}
\usage{
ray(
  betamz,
  umz,
  lat,
  x0 = -130 + 360,
  y0 = -17,
  K = 3,
  dt = 12 * 60 * 60,
  itime = 10,
  direction = -1,
  tl = 1,
  a = 6371000,
  verbose = FALSE,
  ofile
)
}
\arguments{
\item{betamz}{matrix of (longitude = rows x latitude from minor
to major = columns) meridional gradient of the absolute vorticity
in Mercator coordinates. This package includes the
function \code{\link{betaks}} that import NetCDF files
using \strong{ncdf4}.}

\item{umz}{matrix of (longitude = rows x latitude from
minor to major = columns) zonal wind in mercator coordinates.
This package includes the  function \code{\link{betaks}}
that import NetCDF files using \strong{ncdf4}.}

\item{lat}{Numeric vector of latitudes from minor to major
(ex: -90 to 90)}

\item{x0}{Numeric value; longitude (0 to 360 degrees)}

\item{y0}{Numeric value; latitude}

\item{K}{Numeric value; Total Rossby wavenumber}

\item{dt}{Numeric value; Timestep for integration (seconds)}

\item{itime}{Numeric value; total integration time}

\item{direction}{Numeric value (possibilities: 1 or -1).
It controls the wave displacement:
If 1, the wave goes to the north of the source;
If -1, the wave goes to the south of the source.}

\item{tl}{Numeric value; Turning latitude. Do not change this!
It will always start with a positive tl (1) and automatically
change to negative (-1) after the turning latitude.}

\item{a}{Earth's radio (m)}

\item{verbose}{Boolean; if TRUE (default) return messages
during compilation}

\item{ofile}{Character;  Output file name with .csv extension,
for instance, "/user/ray.csv"}
}
\value{
Rossby wave ray paths
}
\description{
This function calculates the Rossby waves ray paths based on
Hoskin and Ambrizzi (1993).
}
\examples{
{
# For Magana & Ambrizzi (1998):
input <- system.file("extdata",
                     "uwnd.mon.mean_300hPa_1997-98DJF.nc",
                      package = "ray")
b <- betaks(ifile = input)
a <- ray(betamz = colMeans(b$betam, na.rm = TRUE),
                umz = colMeans(b$um, na.rm = TRUE),
                lat = b$lat,
                K = 3,
                itime = 10,
                x0 = -130 + 360,
                y0 = -17,
                dt = 12 * 60 * 60,
                direction = -1)

# For Coelho et al. (2015):
input <- system.file("extdata",
                     "uwnd.mon.mean_200hPa_2014JFM.nc",
                      package = "ray")
b <- betaks(ifile = input)
a <- ray(betamz = colMeans(b$betam, na.rm = TRUE),
                umz = colMeans(b$um, na.rm = TRUE),
                lat = b$lat,
                K = 3,
                itime = 30,
                x0 = -135 + 360,
                y0 = -30,
                dt = 6 * 60 * 60,
                direction = -1)

# Simple plot:
a$x2 <- a$x0 - 360
library(ggplot2)
ww <- map_data('world', ylim=c(-60,60))

ggplot() +  theme_bw() +
  scale_y_continuous(limits = c(-60,60)) +
  geom_polygon(data = ww,
               aes(x = long, y = lat, group = group),
               alpha = 0.0, col = "grey") +
  geom_point(data = a[!(a$tun_y0 == -1 |
                          a$tun_y1 == -1 | a$id == 0), ],
             aes(x = x2, y = y0), size = 3) +
  geom_line(data = a[!(a$tun_y0 == -1 |
                         a$tun_y1 == -1 | a$id == 0), ],
            aes(x = x2, y = y0)) +
  ggtitle("") +
  ylab(NULL) + xlab(NULL)
}
}
