% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ray_path.R
\name{ray_path}
\alias{ray_path}
\title{Calculate the ray paths / great circles}
\usage{
ray_path(x, y)
}
\arguments{
\item{x}{vector with the longitude obtained with \code{ray} or \code{ray_source}}

\item{y}{vector with the latitude obtained with \code{ray} or \code{ray_source}}
}
\value{
sf data.frame
}
\description{
This function calculates the great circles using the (x0, y0) obtained
with \code{ray} or \code{ray_source}. It returns a LINESTRING geometry that is ready
for plot.
}
\examples{
{
# For Magana and Ambrizzi (1998):
input <- system.file("extdata",
                     "uwnd.mon.mean_300hPa_1997-98DJF.nc",
                      package = "raytracing")
b <- betaks(u = input)
ma98 <- ray(betam = b$betam,
            u = b$u,
            lat = b$lat,
            K = 3,
            itime = 10,
            x0 = -130,
            y0 = -17,
            dt = 12 * 60 * 60,
            direction = -1)
plot(ma98$geometry, axes = T)
rp <- ray_path(x = ma98$lon, y = ma98$lat)
plot(rp, add = TRUE)

# For Coelho et al. (2015):
input <- system.file("extdata",
                     "uwnd.mon.mean_200hPa_2014JFM.nc",
                      package = "raytracing")
b <- betaks(u = input)
co2015 <- ray(betam = b$betam,
              u = b$u,
              lat = b$lat,
              K = 3,
              itime = 30,
              x0 = -135,
              y0 = -30,
              dt = 6 * 60 * 60,
              direction = -1)
plot(co2015$geometry, axes = T)
rp <- ray_path(x = co2015$lon, y = co2015$lat)
plot(rp, add = TRUE)
}
}
